backend/requirements.txt
sentry-sdk==1.9.8
boto3==1.24.73
botocore==1.27.73
psycopg2-binary==2.9.3
pillow==9.2.0